Move to extension class
