+++ src/ProgressOnderwijsUtils/Data/SqlCommandDebugStringifier.cs
@@ -6,6 +6,7 @@ using System.Globalization;
 using System.Linq;
 using ExpressionToCodeLib;
 using JetBrains.Annotations;
+using ProgressOnderwijsUtils.Extensions;
 
 namespace ProgressOnderwijsUtils
 {
