+++ src/ProgressOnderwijsUtils/Extensions/ExceptionExtensions.cs
@@ -1,7 +1,9 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Data.SqlClient;
 using System.Linq;
 using System.Text;
+using JetBrains.Annotations;
 
 namespace ProgressOnderwijsUtils.Extensions
 {
@@ -16,6 +18,20 @@ namespace ProgressOnderwijsUtils.Extensions
             } else {
                 return exception.InnerException is Exception child && AnyNestingLevelMatches(child, predicate);
             }
+        }
+
+        public static bool IsSqlTimeoutException([CanBeNull] this Exception e)
+        {
+            if (e is AggregateException aggregateException) {
+                return aggregateException.InnerExceptions.All(IsSqlTimeoutException);
+            }
+
+            for (var current = e; current != null; current = current.InnerException) {
+                if (current is SqlException sqlE && sqlE.Number == -2) {
+                    return true;
+                }
+            }
+            return false;
         }
     }
 }
